#1015 tidy flyscan xraycentre tests
